backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy[asyncio]>=2.0.25,<3.0.0
aiosqlite>=0.17.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0
httpx>=0.26.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
praw>=7.7.0,<8.0.0
asyncpraw>=7.7.0,<8.0.0
spacy>=3.7.0,<4.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.4.0,<2.0.0
loguru>=0.7.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
email-validator>=2.1.0,<3.0.0
vaderSentiment>=3.3.2,<4.0.0
